requirements.txt
pydantic-ai
python-dotenv
python-telegram-bot
# For Google Calendar MCP support if needed directly
# mcp 
